pyproject.toml
[project]
name = "fastapilearn"
version = "0.1.0"
description = "Learning fastapi"
authors = [
    {name = "Karthik"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.115.8,<0.116.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "pydantic-settings (>=2.8.0,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "psycopg[binary] (>=3.2.5,<4.0.0)",
    "sqlmodel (>=0.0.22,<0.0.23)",
    "sqlalchemy[asyncio] (>=2.0.38,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
